requirements.txt
flask
opencv-python
numpy
faiss-cpu
tqdm
inference-sdk
werkzeug
webdriver-manager
selenium
requests